Extremely speed-up puzzle resolution
This is based on the observation that there is no point in exploring
the rest of the board if an empty spot could not be filled right now,
because it will need to be filled later regardless. So any subsequent
search after skipping an empty spot will just lead back to previously
explored states.